working on the shader-reader

+++ MyShaderAnalysis/vcsparsing/ShaderDataBlock.cs
@@ -4,8 +4,8 @@ namespace ValveResourceFormat.CompiledShader
     public abstract class ShaderDataBlock
     {
         public ShaderDataReader datareader { get; }
-        protected int start { get; }
-        protected ShaderDataBlock(ShaderDataReader datareader, int offsetAtStartOfBlock)
+        protected long start { get; }
+        protected ShaderDataBlock(ShaderDataReader datareader, long offsetAtStartOfBlock)
         {
             this.start = offsetAtStartOfBlock;
             this.datareader = datareader;
